Split out Werror
